+++ src/app/app.module.ts
@@ -36,8 +36,6 @@ import { MainComponent } from './main/main.component';
 import { HomeComponent } from './home/home.component';
 import { CreateGroupComponent } from './create-group/create-group.component';
 import { ChatListItemComponent } from './chat-list-item/chat-list-item.component';
-import { MustMatchDirective } from './shared/must-match.directive';
-import { MustNotMatchDirective } from './shared/must-not-match.directive';
 
 @NgModule({
   declarations: [
@@ -54,8 +52,6 @@ import { MustNotMatchDirective } from './shared/must-not-match.directive';
     HomeComponent,
     CreateGroupComponent,
     ChatListItemComponent,
-    MustMatchDirective,
-    MustNotMatchDirective,
   ],
   imports: [
     BrowserModule,
