--- src/app/shared/must-match.directive.spec.ts
@@ -1,8 +1,0 @@
-import { MustMatchDirective } from './must-match.directive';
-
-describe('MustMatchDirective', () => {
-  it('should create an instance', () => {
-    const directive = new MustMatchDirective();
-    expect(directive).toBeTruthy();
-  });
-});
