--- src/app/shared/must-not-match.directive.spec.ts
@@ -1,8 +1,0 @@
-import { MustNotMatchDirective } from './must-not-match.directive';
-
-describe('MustNotMatchDirective', () => {
-  it('should create an instance', () => {
-    const directive = new MustNotMatchDirective();
-    expect(directive).toBeTruthy();
-  });
-});
